Separate chat code from server code
